+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 ## [Unreleased]
 
+
+## [2.0.11] - 2025-09-18
+
+### Added
+
+- NEW FEATURE: SetUp Flutter intl: right click any file in Project View, and select "Setup Flutter Intl" if required
+- Partially: Import / created statements already use l10n.yaml file instead of settings - refactor is not entirely complete, though
+- Ask on project startup for set-up (once)
+
+
 ## [2.0.10] - 2025-09-18
 
 ### Added
@@ -202,7 +212,8 @@
 
 - initial release
 
-[Unreleased]: https://github.com/keeyzar/gpt-helper/compare/v2.0.9...HEAD
+[Unreleased]: https://github.com/keeyzar/gpt-helper/compare/v2.0.10...HEAD
+[2.0.10]: https://github.com/keeyzar/gpt-helper/compare/v2.0.9...v2.0.10
 [2.0.9]: https://github.com/keeyzar/gpt-helper/compare/v2.0.8...v2.0.9
 [2.0.8]: https://github.com/keeyzar/gpt-helper/compare/v2.0.7...v2.0.8
 [2.0.7]: https://github.com/keeyzar/gpt-helper/compare/v2.0.6...v2.0.7